Update Table actions and device type option text
